x/geotrace/types/message_add_location_proof_test.go
package types

import (
	"testing"

	"geotrace/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgAddLocationProof_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgAddLocationProof
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgAddLocationProof{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgAddLocationProof{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
